add anomaly manager
