+++ Assets/Scripts/Managers/AnomalyManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6dfb4e68e8e374175bd1f71b9b15275c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
